Created the linked list structure for javascript. Created the structure for the project list, employee list and action list and their respective nodes.
Implemented the ProgrammerMacro.js decision tree and the action generated. However the tree leafs are not implemented.
Created a script BehaivorPlanner.js which contains the information necessary for all decisions tree. All employees have this script.
Created a prototype window to show the action for employee #01.

Removed a warning in SaveGame.js (repeated library)
Changed Working.js so it can support the changes in ProgrammerMacro.js