Changed PapelWindow.js size, increasing it to be easier to click on the close button.
Implemented the architect's decision tree and because of that, it was necessary to change all report-related scripts to break the architect bug score to system and integration scores. Scripts affected: WeeklyReport.js, Equipe.js, Working.js, ArchitectMacro.js, FuncWindow.js.

Expanded ActionList.js information on the nodes, to include "influenced" and "who", as in the name of the person who did the action.
Changed the coordinations of the flying text 1 and 2 to be in correct order in FloatingLines.js

Changed the PopUpListGUI.js to remove the restriction in the GUI when is in macro or micro mode.

Don't remember what i changed in WindowController.js ... probably i removed some commentary... >.>

Changed WorkingHoursWindow.js to allow the player to set the tasks for each role, as of now only programmer and architect have entry in the window. That window will change the variables in BehaviorPlanner.js for the employee.

Created the script Influence.js that will record all actions that can affect other roles. As of now, only the architect writes in it and the programmer read from it.

With Influence.js and the influence variable in ActionNodes, now it is possible to follow the paths from the LogWindow.js to reach the action that influenced the outcome. The difference is that it opens a window showing all data in Influence.js

I changed the programmer's decision tree to be easier to read and understand, and the leaf functions to use the Influence.js script.

End of log.