Reviewed the Analysts's decision tree in Decision Tree.xlsx
Created the analyst decision tree on AnalystMacro.js
Because of that, some changes were made on BehaivorPlanner.js, ProgrammerMacro.js, Project.js, IngameMenu.js, Working.js and WorkingHoursWindow.js

Disabled the Gameplay selection window for now. It is always set to be on Macro mode.
Removed the distinction of Macro and Micro when employees work.

Added in LogWindow.js the option to walk in the action list.
Added as well the option to see the project details.