Created a script named ManagerPlanner.js that will control some of the actions made by the manager.
Changed in AnalystMacro.js and ArchitectMacro.js the work "Rounded" to "Balanced" on the reports. As well as in WorkingHoursWindow.js
Corrected some issues with TesterMacro.js and Working.js about the Tester.
Corrected an old bug from Training both in Dialog.js, PopUpListGUI.js and EspWindow.js. Now the Lock (cant change roles nor train again while training) is correctly applied.

Increased the delay for the Secondary roles before they are processed. Changed GameTime.js to compensate (although the only timespeed changed was the one not being used)

Dont remember what i did on HireWindow.js ...
Implemented a new function on Funcionario.js called SetRates()

On ManagerPlanner.js has a new Window for Developer Usage to test the manager actions. As such IngameMenu.js was changed to allow the access of the window.

Cleaned some garbage on Equipe.js and move others to ManagerPlanner.js

Created some functions on ManagerPlanner.js to allow the distribution of roles for each employee, checking their best 2 roles. If no employee has the role in their list of best and second best, them we try by role substitution by priority. Function named NeedRole()

Created a dummy function named AnalysisMode() to test NeedRole() for when a certain Mode is selected (in this case, Analysis). So employees are deployed according to the Mode's priorities. In this case need to have at least 1 main Analyst and 1 secondary Analyst. The rest is deployed as Programmers, Archtect and Tester.