Camera ready version 3 from paper Provenance in Games Copyright for SBGames 2012
SDM: 
Added a function to allow dialog through mouse pointer and click.
Moved from the prototype version to the first version of Influence, adding analyst, architect, programmer, tester sections and those that provide influences.
PS: There is a bug related to manager influences and architect-programmer. Also need to correct the bug on list movement for influence, when there are more than 1 person with the same role (ex 2 architects aiding programmers). For some reason, it is not reconizing elements from the list, only the extremes.